requirements.txt
click==7.1.2
Flask==1.1.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
gevent==23.9.1
grpcio==1.59.0
h5py==3.10.0
Pillow==10.0.1
scikit-learn==1.3.1
tensorflow==2.12.0
